made a lot of improvements
seems like i can successfully generate and read courses and files. json is working. still need to enable master file reading and writing.

+++ CourseInfo.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from functools import total_ordering
 from pathlib import Path
 from copy import deepcopy
 import json
@@ -14,6 +15,7 @@ def get_course_directory(identifier: str) -> Path:
     return COURSE_DIR / identifier
 
 
+@total_ordering
 @dataclass()
 class CourseInfo:
     year: int
@@ -26,16 +28,25 @@ class CourseInfo:
     website: str
 
     def __post_init__(self):
-        self.identifier = f'{self.department[:4].upper()}{self.number}{self.institution[0].upper}'
-        self.syllabus = Syllabus()
+        self.identifier = f'{self.department[:4].upper()}{self.number}{self.institution[0].upper()}'
 
-        get_course_directory(self.identifier).mkdir()
-        get_json_location(self.identifier).touch()
+        self.directory = get_course_directory(self.identifier)
+        self.json_location = get_json_location(self.identifier)
+        self.write_json()
 
-    def write_json_file(self):
+    def get_syllabus(self):
+        return self.directory / 'syllabus.pdf'
+
+    def write_json(self):
+        if self.json_location.exists():
+            return
+
+        if not self.json_location.parent.exists():
+            self.json_location.parent.mkdir(parents=True)
+
         with open(get_json_location(self.identifier), 'w') as file:
             json_vars = deepcopy(vars(self))
-            for k in ('identifier', 'syllabus'):
+            for k in ('identifier', 'directory', 'json_location'):
                 json_vars.pop(k, None)
             json.dump(json_vars, file)
 
@@ -45,28 +56,11 @@ class CourseInfo:
     def __eq__(self, other):
         return self.ct() == other.ct()
 
-    def __ne__(self, other):
-        return self.ct() != other.ct()
-
     def __lt__(self, other):
         return self.ct() < other.ct()
 
-    def __le__(self, other):
-        return self.ct() <= other.ct()
-
-    def __gt__(self, other):
-        return self.ct() > other.ct()
-
-    def __ge__(self, other):
-        return self.ct() >= other.ct()
-
-    def __cmp__(self, other):
-        return self.ct().__cmp__(other.ct())
-
-
-class Syllabus:
-    def __get__(self, owner: CourseInfo, objtype=None):
-        return Path(owner.directory / 'syllabus.pdf')
+    def __str__(self):
+        return f'{self.identifier[:-1]}: {self.title}'
 
 
 def CourseInfo_from_json(file_path: Path or str):
@@ -80,3 +74,11 @@ def CourseInfo_from_json(file_path: Path or str):
 def CourseInfo_from_identifier(identifier: str) -> CourseInfo:
     file_path = get_json_location(identifier)
     return CourseInfo_from_json(file_path)
+
+
+if __name__ == "__main__":
+    ci = CourseInfo(year=2021, semester='Fall', department='COMP', number=332,
+                    title='Analysis of Algorithms', institution='Dickinson', professor='Richard Forrester', website='')
+    print(ci)
+    ci2 = CourseInfo_from_json(ci.json_location)
+    print(ci2)
